pyproject.toml
[tool.poetry]
name = "ai-report-manager"
version = "0.1.0"
description = ""
authors = ["Edoardo Tantari"]
readme = "README.md"
packages = [{include = "ai_report_manager"}]

[tool.poetry.dependencies]
python = "^3.12"
marimo = "^0.14.9"
langchain-core = "^0.3.66"
langchain-mistralai = "^0.2.10"
python-dotenv = "^1.1.1"
langchain = "^0.3.26"
langgraph = "^0.5.0"
sounddevice = "^0.5.2"
soundfile = "^0.13.1"
scipy = "^1.16.0"
wavio = "^0.0.9"
transformers = "^4.53.0"
torch = "^2.7.1"
tensorflow = "^2.19.0"
tf-keras = "^2.19.0"
utils = "^1.0.2"
optional = "^0.0.1"
ipykernel = "^6.29.5"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
